+++ src/scene/Scene.tsx
@@ -26,6 +26,7 @@ import { DragController } from './DragController'
 import { deselectOnMiss } from './deselectOnMiss'
 import { Effects } from './Effects'
 import { FinishDropSurface } from './FinishDropSurface'
+import { FinishEyedropperSurface } from './FinishEyedropperSurface'
 import { FrameRenderedNotifier } from './FrameRenderedNotifier'
 import { GridOverlay } from './GridOverlay'
 import { HqRenderController } from './HqRenderController'
@@ -140,6 +141,7 @@ export function Scene() {
         <PlacementDropAnimator />
         <DragController />
         <FinishDropSurface />
+        <FinishEyedropperSurface />
         <MarqueeCameraTracker />
         <CameraRig />
         <CameraForwardTracker />